Refactoring of falcon keygen and gpv

+++ src/utils/arith/falcon.h
@@ -385,8 +385,8 @@ int falcon_keygen_make(falcon_keygen *fk, int comp,
 
 
 
-int solve_NTRU(falcon_keygen *fk, int16_t *F, int16_t *G,
-	const int16_t *f, const int16_t *g);
+int solve_NTRU(falcon_keygen *fk, int32_t *F, int32_t *G,
+	const int32_t *f, const int32_t *g);
 
 #ifdef __cplusplus
 }
